New auxiliary File for Bonsai Preprocessing
