+++ tf/bonsai/bonsaiPreProcess.py
@@ -1,0 +1,140 @@
+import argparse
+import numpy as np
+
+# Functions to check sanity of input arguments
+
+
+def checkIntPos(value):
+    ivalue = int(value)
+    if ivalue <= 0:
+        raise argparse.ArgumentTypeError(
+            "%s is an invalid positive int value" % value)
+    return ivalue
+
+
+def checkIntNneg(value):
+    ivalue = int(value)
+    if ivalue < 0:
+        raise argparse.ArgumentTypeError(
+            "%s is an invalid non-neg int value" % value)
+    return ivalue
+
+
+def checkFloatNneg(value):
+    fvalue = float(value)
+    if fvalue < 0:
+        raise argparse.ArgumentTypeError(
+            "%s is an invalid non-neg float value" % value)
+    return fvalue
+
+
+def checkFloatPos(value):
+    fvalue = float(value)
+    if fvalue <= 0:
+        raise argparse.ArgumentTypeError(
+            "%s is an invalid positive float value" % value)
+    return fvalue
+
+
+def getArgs():
+    '''
+    Function to parse arguments
+    '''
+    parser = argparse.ArgumentParser(
+        description='HyperParams for Bonsai Algorithm')
+    parser.add_argument('-dir', '--data_dir', required=True,
+                        help='Data directory containing train.npy and test.npy')
+
+    parser.add_argument('-d', '--depth', type=checkIntNneg, default=2,
+                        help='Depth of Bonsai Tree (default: 2 try: [0, 1, 3])')
+    parser.add_argument('-p', '--projDim', type=checkIntPos, default=10,
+                        help='Projection Dimension (default: 20 try: [5, 10, 30])')
+    parser.add_argument('-s', '--sigma', type=float, default=1.0,
+                        help='Parameter for sigmoid sharpness (default: 1.0 try: [3.0, 0.05, 0.1]')
+    parser.add_argument('-e', '--epochs', type=checkIntPos, default=42,
+                        help='Total Epochs (default: 42 try:[100, 150, 60])')
+    parser.add_argument('-b', '--batchSize', type=checkIntPos,
+                        help='Batch Size to be used (default: max(100, sqrt(train_samples)))')
+    parser.add_argument('-lr', '--learningRate', type=checkFloatPos, default=0.01,
+                        help='Initial Learning rate for Adam Oprimizer (default: 0.01)')
+
+    parser.add_argument('-rW', type=float, default=0.0001,
+                        help='Regularizer for predictor parameter W  (default: 0.0001 try: [0.01, 0.001, 0.00001])')
+    parser.add_argument('-rV', type=float, default=0.0001,
+                        help='Regularizer for predictor parameter V  (default: 0.0001 try: [0.01, 0.001, 0.00001])')
+    parser.add_argument('-rT', type=float, default=0.0001,
+                        help='Regularizer for branching parameter Theta  (default: 0.0001 try: [0.01, 0.001, 0.00001])')
+    parser.add_argument('-rZ', type=float, default=0.00001,
+                        help='Regularizer for projection parameter Z  (default: 0.00001 try: [0.001, 0.0001, 0.000001])')
+
+    parser.add_argument('-sW', type=checkFloatPos,
+                        help='Sparsity for predictor parameter W  (default: For Binary classification 1.0 else 0.2 try: [0.1, 0.3, 0.5])')
+    parser.add_argument('-sV', type=checkFloatPos,
+                        help='Sparsity for predictor parameter V  (default: For Binary classification 1.0 else 0.2 try: [0.1, 0.3, 0.5])')
+    parser.add_argument('-sT', type=checkFloatPos,
+                        help='Sparsity for branching parameter Theta  (default: For Binary classification 1.0 else 0.2 try: [0.1, 0.3, 0.5])')
+    parser.add_argument('-sZ', type=checkFloatPos, default=0.2,
+                        help='Sparsity for projection parameter Z  (default: 0.2 try: [0.1, 0.3, 0.5])')
+
+    return parser.parse_args()
+
+
+def preProcessData(data_dir):
+    '''
+    Function to pre-process input data
+    Expects a .npy file of form [lbl feats] for each datapoint
+    Outputs a train and test set datapoints appended with 1 for Bias induction
+    dataDimension, numClasses are inferred directly
+    '''
+    train = np.load(data_dir + '/train.npy')
+    test = np.load(data_dir + '/test.npy')
+
+    dataDimension = int(train.shape[1]) - 1
+
+    Xtrain = train[:, 1:dataDimension + 1]
+    Ytrain_ = train[:, 0]
+    numClasses = max(Ytrain_) - min(Ytrain_) + 1
+
+    Xtest = test[:, 1:dataDimension + 1]
+    Ytest_ = test[:, 0]
+
+    numClasses = int(max(numClasses, max(Ytest_) - min(Ytest_) + 1))
+
+    # Mean Var Normalisation
+    mean = np.mean(Xtrain, 0)
+    std = np.std(Xtrain, 0)
+    std[std[:] < 0.000001] = 1
+    Xtrain = (Xtrain - mean) / std
+
+    Xtest = (Xtest - mean) / std
+    # End Mean Var normalisation
+
+    lab = Ytrain_.astype('uint8')
+    lab = np.array(lab) - min(lab)
+
+    lab_ = np.zeros((Xtrain.shape[0], numClasses))
+    lab_[np.arange(Xtrain.shape[0]), lab] = 1
+    if (numClasses == 2):
+        Ytrain = np.reshape(lab, [-1, 1])
+    else:
+        Ytrain = lab_
+
+    lab = Ytest_.astype('uint8')
+    lab = np.array(lab) - min(lab)
+
+    lab_ = np.zeros((Xtest.shape[0], numClasses))
+    lab_[np.arange(Xtest.shape[0]), lab] = 1
+    if (numClasses == 2):
+        Ytest = np.reshape(lab, [-1, 1])
+    else:
+        Ytest = lab_
+
+    trainBias = np.ones([Xtrain.shape[0], 1])
+    Xtrain = np.append(Xtrain, trainBias, axis=1)
+    testBias = np.ones([Xtest.shape[0], 1])
+    Xtest = np.append(Xtest, testBias, axis=1)
+
+    if numClasses == 2:
+        numClasses = 1
+
+    return dataDimension + 1, numClasses, Xtrain, Ytrain, Xtest, Ytest
